update test scripts
